# Python 3.7 compatibility due to new requirements to re.sub repl.